reshuffle Stores implement update for TxHistory Store
