+++ lib/core/cardano-wallet-core.cabal
@@ -195,12 +195,15 @@ library
       Cardano.Wallet.CoinSelection.Internal.Context
       Cardano.Wallet.Compat
       Cardano.Wallet.DB
-      Cardano.Wallet.DB.Checkpoints
       Cardano.Wallet.DB.MVar
       Cardano.Wallet.DB.Model
       Cardano.Wallet.DB.Sqlite
       Cardano.Wallet.DB.Sqlite.AddressBook
-      Cardano.Wallet.DB.Sqlite.CheckpointsOld
+      Cardano.Wallet.DB.Stores.Checkpoints.Model
+      Cardano.Wallet.DB.Stores.Checkpoints.Sqlite
+      Cardano.Wallet.DB.Stores.TxHistory.Model
+      Cardano.Wallet.DB.Stores.TxHistory.Sqlite
+      Cardano.Wallet.DB.Stores.TxHistory.Update
       Cardano.Wallet.DB.Sqlite.Migration
       Cardano.Wallet.DB.Sqlite.Stores
       Cardano.Wallet.DB.Sqlite.TH
