+++ lib/core/src/Cardano/Wallet/DB/Stores/Checkpoints/Sqlite.hs
@@ -29,13 +29,10 @@
 -- * Inline the contents of this module into its new name
 --   "Cardano.Wallet.DB.Sqlite.Stores"
 
-module Cardano.Wallet.DB.Sqlite.CheckpointsOld
-    ( mkStoreWallets
+module Cardano.Wallet.DB.Stores.Checkpoints.Sqlite
+    ( mkStoreCheckpoints
     , PersistAddressBook (..)
     , blockHeaderFromEntity
-
-    -- * Testing
-    , mkStoreWallet
     )
     where
 
@@ -49,8 +46,6 @@ import Cardano.DB.Sqlite
     ( dbChunked )
 import Cardano.Wallet.DB
     ( ErrBadFormat (..) )
-import Cardano.Wallet.DB.Checkpoints
-    ( DeltaCheckpoints (..), loadCheckpoints )
 import Cardano.Wallet.DB.Sqlite.AddressBook
     ( AddressBookIso (..)
     , Discoveries (..)
@@ -71,7 +66,6 @@ import Cardano.Wallet.DB.Sqlite.TH
     , SharedState (..)
     , UTxO (..)
     , UTxOToken (..)
-    , Wallet (..)
     )
 import Cardano.Wallet.DB.Sqlite.Types
     ( BlockId (..)
@@ -81,14 +75,8 @@ import Cardano.Wallet.DB.Sqlite.Types
     , hashOfNoParent
     , toMaybeHash
     )
-import Cardano.Wallet.DB.WalletState
-    ( DeltaMap (..)
-    , DeltaWalletState
-    , DeltaWalletState1 (..)
-    , WalletCheckpoint (..)
-    , WalletState (..)
-    , getSlot
-    )
+import Cardano.Wallet.DB.Stores.Checkpoints.Model
+    ( DeltaCheckpoints (..), WalletCheckpoint (..), getSlot, loadCheckpoints )
 import Cardano.Wallet.Primitive.AddressDerivation
     ( Depth (..)
     , MkKeyFingerprint (..)
@@ -116,8 +104,6 @@ import Control.Monad.IO.Class
     ( MonadIO (..) )
 import Control.Monad.Trans.Class
     ( lift )
-import Control.Monad.Trans.Except
-    ( ExceptT (..), runExceptT )
 import Control.Monad.Trans.Maybe
     ( MaybeT (..) )
 import Control.Monad.Trans.Reader
@@ -129,7 +115,7 @@ import Data.DBVar
 import Data.Functor
     ( (<&>) )
 import Data.Generics.Internal.VL.Lens
-    ( view, (^.) )
+    ( (^.) )
 import Data.Map.Strict
     ( Map )
 import Data.Maybe
@@ -180,70 +166,6 @@ import qualified Control.Monad.STM as STM
 import qualified Data.Map.Merge.Strict as Map
 import qualified Data.Map.Strict as Map
 
-{-------------------------------------------------------------------------------
-    WalletState Store
--------------------------------------------------------------------------------}
--- | Store for 'WalletState' of multiple different wallets.
-mkStoreWallets
-    :: forall s key. (PersistAddressBook s, key ~ W.WalletId)
-    => Store (SqlPersistT IO)
-        (DeltaMap key (DeltaWalletState s))
-mkStoreWallets = Store{loadS=load,writeS=write,updateS=update}
-  where
-    write = error "mkStoreWalletsCheckpoints: not implemented"
-
-    update _ (Insert wid a) =
-        writeS (mkStoreWallet wid) a
-    update _ (Delete wid) = do
-        -- FIXME LATER during ADP-1043:
-        --  Deleting an entry in the Checkpoint table
-        --  will trigger a delete cascade. We want this cascade
-        --  to be explicit in our code.
-        deleteWhere [CheckpointWalletId ==. wid]
-    update _ (Adjust wid da) =
-        updateS (mkStoreWallet wid) undefined da
-        -- FIXME LATER during ADP-1043:
-        --   Remove 'undefined'.
-        --   Probably needs a change to 'Data.DBVar.updateS'
-        --   to take a 'Maybe a' as parameter instead of an 'a'.
-
-    load = do
-        wids <- fmap (view #walId . entityVal) <$> selectAll
-        runExceptT $ do
-            xs <- forM wids $ ExceptT . loadS . mkStoreWallet
-            pure $ Map.fromList (zip wids xs)
-      where
-        selectAll :: SqlPersistT IO [Entity Wallet]
-        selectAll = selectList [] []
-
--- | Store for 'WalletState' of a single wallet.
-mkStoreWallet
-    :: forall s. PersistAddressBook s
-    => W.WalletId
-    -> Store (SqlPersistT IO) (DeltaWalletState s)
-mkStoreWallet wid =
-    Store{ loadS = load, writeS = write, updateS = \_ -> update }
-  where
-    storeCheckpoints = mkStoreCheckpoints wid
-
-    load = do
-        eprologue <- maybe (Left $ toException ErrBadFormatAddressPrologue) Right
-            <$> loadPrologue wid
-        echeckpoints <- loadS storeCheckpoints
-        pure $ WalletState <$> eprologue <*> echeckpoints
-
-    write wallet = do
-        insertPrologue wid (wallet ^. #prologue)
-        writeS storeCheckpoints (wallet ^. #checkpoints)
-
-    update =
-         -- first update in list is last to be applied!
-        mapM_ update1 . reverse
-    update1 (ReplacePrologue prologue) =
-        insertPrologue wid prologue
-    update1 (UpdateCheckpoints delta) =
-        -- FIXME LATER during ADP-1043: remove 'undefined'
-        updateS storeCheckpoints undefined delta
 
 -- | Store for the 'Checkpoints' belonging to a 'WalletState'.
 mkStoreCheckpoints
